Add cxx11 string ABI probing to BB tarball selection
Use this to automatically fail out when users try to use a `gcc4`
tarball with a system that compiles for `cxx11`.  As of the time
of this commit, this will apply to LLVM, OpenBLAS, OpenLibm and
SuiteSparse, although the only place we truly care aobut this is
LLVM, because it's the only one that passes around `std::string`s.

Also reduce the amount of work we're forcing `make` to do when
constructing BB triplets